Fixed possible compatibility issue
This allows the program to read in first what the user previously
selected as the library location, as opposed to making the user set the
location every time he/she opens the program. It also prevents an issue
to occur where if compiled and the user does not select a location, an
IOError would arise. This prevents that.